update code rename nowplaying to clementine
